+++ push/lib-push/src/main/java/ru/surfstudio/android/notification/service/BaseAppStartReceiver.kt
@@ -3,15 +3,17 @@ package ru.surfstudio.android.notification.service
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import ru.surfstudio.android.logger.Logger
 
 /**
- * Базовый ресивер куда приходит ивент о закрытии сервиса [NotificationReviverService]
+ * Базовый ресивер, куда приходит ивент о закрытии сервиса [NotificationReviverService]
  */
-abstract class BaseAppStartReceiver<S : BaseNotificationProcessStarter>: BroadcastReceiver() {
+abstract class BaseAppStartReceiver: BroadcastReceiver() {
 
-    abstract fun getNotificationProcessStarter(): S
+    abstract fun getNotificationProcessStarter(): BaseNotificationProcessStarter
 
     override fun onReceive(context: Context?, intent: Intent?) {
+        Logger.d("AppStartReceiver :: onReceive")
         getNotificationProcessStarter().startNotificationServiceReviver()
     }
 }