+++ push/lib-push/src/main/java/ru/surfstudio/android/notification/service/NotificationProcessStarterHolder.kt
@@ -2,7 +2,9 @@ package ru.surfstudio.android.notification.service
 
 /**
  * Холдер-класс со ссылкой на конкретный
- * [BaseNotificationProcessStarter]
+ * [BaseNotificationProcessStarter];
+ * сервис из push-модуля не знает о конкретной реализации AppComponent,
+ * откуда он может взять ссылку
  */
 object NotificationProcessStarterHolder {