+++ template/base_feature/src/main/java/ru/surfstudio/standard/application/notification/NotificationProcessStarter.kt
@@ -2,13 +2,17 @@ package ru.surfstudio.standard.application.notification
 
 import android.content.Context
 import ru.surfstudio.android.notification.service.BaseNotificationProcessStarter
-import ru.surfstudio.android.notification.ui.UiHandler
+import ru.surfstudio.android.notification.service.NotificationReviverService
 import javax.inject.Inject
 
+/**
+ * Целевой класс для запуска [NotificationReviverService],
+ * который предоставляет имя класса [AppStartReceiver] для запуска
+ */
 class NotificationProcessStarter @Inject constructor(
-        uiHandler: UiHandler,
         applicationContext: Context
-) : BaseNotificationProcessStarter(uiHandler, applicationContext) {
+) : BaseNotificationProcessStarter(applicationContext) {
 
-    override fun getAppStartReceiverClassName() = AppStartReceiver::class.java.canonicalName
+    override fun getAppStartReceiverClassName(): String 
+            = AppStartReceiver::class.java.canonicalName as String
 }